refactor controllers to move them under Resources namespace for better organization
